requirements.txt
notion-client
python-dotenv
pytest
ruff
openai
sentence-transformers
qdrant-client
scikit-learn
numpy
notion2md 